folder: folder.js -> folder-data.js |  folder-create.js
